chore : 매칭페이지 보호자 버전 반응형 수정
