+++ back/build.gradle
@@ -24,13 +24,13 @@ repositories {
 }
 
 dependencies {
+    //필요한 의존성들을 묶어서 제공하는 스타터 패키지입니다.
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
     //Spring Boot에서 메일 전송 기능을 간편하게 구현할 수 있도록
-    //필요한 의존성들을 묶어서 제공하는 스타터 패키지입니다.
     implementation 'org.springframework.boot:spring-boot-starter-mail'
     //메일형식을 html로 쉽게 작성하기위한 템플릿엔진
     implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
