+++ front/src/components/Paging.jsx
@@ -3,6 +3,8 @@ import styled from 'styled-components';
 import { Pagination, Stack } from '@mui/material';
 
 const Paging = ({ totalPage, currentPage, chagneCurrentPage }) => {
+
+
   return (
     <>
       <PaginationWrapper>
