+++ front/src/components/PatientCardGroup.jsx
@@ -18,20 +18,38 @@ import {
   ReportButton,
   CareLogButton,
   PageWrapper,
+  EmptyMessage,
 } from '../styles/MatchingCard';
 import { ProfileCard } from '../styles/MatchingCard';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
-import { Btn } from '../styles/common/Board';
-import { BtnWrap } from '../styles/PatientRegistration';
+import SearchDate from './SearchDate';
 import { Button } from '../styles/Auth.styles';
+import Paging from './Paging';
+import MatchCareGiverCard from './MatchCareGiverCard';
 
-
-
-const PatientCardGroup = ({ patient, caregiverList, onClose }) => {
+const PatientCardGroup = ({
+  patient,
+  caregiverList,
+  endedCaregiverList,
+  onClose,
+  activeTab,
+  startDate,
+  endDate,
+  handleStartDateChange,
+  handleEndDateChange,
+  handleSearchClick,
+  endedCurrentPage,
+  endedTotalPage,
+  handleEndedPageChange,
+  setShowReviewModal,
+  setSelectedCaregiver,
+  isOpen
+}) => {
   if (!patient) return <p>선택된 환자 정보가 없습니다.</p>;
   const CLOUDFRONT_URL = 'https://d20jnum8mfke0j.cloudfront.net/';
   //이미지 경로 갖고오고 없다면 기본이미지
+
   const getProfileImageUrl = (path, type = 'caregiver') => {
     if (!path) {
       return type === 'patient' ? pat_profileImage : care_profileImage;
@@ -40,83 +58,128 @@ const PatientCardGroup = ({ patient, caregiverList, onClose }) => {
     return `${CLOUDFRONT_URL}${cleanPath}`;
   };
 
+  let currentList;
+
+  if (activeTab === 'matching') {
+    currentList = caregiverList;
+  } else if (activeTab === 'matched') {
+    currentList = endedCaregiverList;
+  }
 
   const navigate = useNavigate();
 
-  return(
-  <div>
-    <ProfileCard type="patient">
-      <ProfileImage
-        src={getProfileImageUrl(patient.profileImage, 'patient')}
-        alt="환자"
-        onError={(e) => {
-          e.target.onerror = null;
-          e.target.src = pat_profileImage;
-        }}
-      />
-      <ProfileInfo>
-        <UserName>{patient.patName} 님</UserName>
-        <UserAge>
-          나이 {patient.patAge}세({patient?.patGender === 'M' ? '남' : '여'})
-        </UserAge>
+  return (
+    <Div>
+      <ProfileCard type="patient">
+        <ProfileImage
+          src={getProfileImageUrl(patient.profileImage, 'patient')}
+          alt="환자"
+          onError={(e) => {
+            e.target.onerror = null;
+            e.target.src = pat_profileImage;
+          }}
+        />
+        <ProfileInfo>
+          <UserName>{patient.patName} 님</UserName>
+          <UserAge>
+            나이 {patient.patAge}세({patient?.patGender === 'M' ? '남' : '여'})
+          </UserAge>
 
-        <InfoButton onClick={() => navigate(`/report/${patient.patNo}`)}>간병일지 보기</InfoButton>
-      </ProfileInfo>
-    </ProfileCard>
+          <InfoButton onClick={() => navigate(`/report/${patient.patNo}`)}>간병일지 보기</InfoButton>
+        </ProfileInfo>
+      </ProfileCard>
 
-    <Button onClick={onClose(null)}>상세보기 닫기</Button>
+      <Button onClick={onClose(null)}>상세보기 닫기</Button>
 
-    {caregiverList && caregiverList.length > 0 ? (
-      caregiverList.map((care) => (
-        <CargiverWrap key={care.caregiverNo}>
-          <CaregiverDiv>
-            <CaregiverImg
-              src={getProfileImageUrl(care.caregiverProfileImage, 'caregiver')}
-              alt="간병인"
-              onError={(e) => {
-                e.target.onerror = null;
-                e.target.src = care_profileImage;
-              }}
-            />
-            <CaregiverTextDiv>
-              <ProfileTextGray>
-                <ProfileTextStrong>{care.userName}</ProfileTextStrong> 님
-              </ProfileTextGray>
-              <ProfileTextGray>
-                나이
-                <ProfileTextStrong>
-                  {care.age} 세 ({care.gender === 'M' ? '남' : care.gender === 'F' ? '여' : '성별 정보 없음'})
-                </ProfileTextStrong>
-              </ProfileTextGray>
-            </CaregiverTextDiv>
-          </CaregiverDiv>
+    {activeTab === 'matched'  ? (
+        <SearchDate
+          startDate={startDate}
+          endDate={endDate}
+          handleStartDateChange={handleStartDateChange}
+          handleEndDateChange={handleEndDateChange}
+          handleSearchClick={handleSearchClick}
+        />
+      ) : (
+        <></>
+      )}
 
-          <CargiverButtonDiv>
-            <CareLogButton
-              onClick={() =>
-                navigate(`/caregiverProfile/${Number(care.caregiverNo)}`, {
-                  state: { matNo: care.matNo },
-                })
-              }
-            >
-              간병인 정보
-            </CareLogButton>
-          </CargiverButtonDiv>
-        </CargiverWrap>
-      ))
-    ) : (
-      <EmptyMessage>매칭된 간병이 없습니다.</EmptyMessage>
-    )}
-  </div>
+      {currentList &&
+        currentList.length > 0 ?(
+        currentList.map((care) => (
+          // 모바일일때는 반응형으로 display none: flex로 잡고
+          // 여기는 모바일일 때 보여야 상세보기로 display 조건을 isOpen 속성을 추가하여 caregiverCard와 다르게 써야함 
+          <NewWrap key={care.caregiverNo}>
+            <CaregiverDiv>
+              <CaregiverImg
+                src={getProfileImageUrl(care.caregiverProfileImage, 'caregiver')}
+                alt="간병인"
+                onError={(e) => {
+                  e.target.onerror = null;
+                  e.target.src = care_profileImage;
+                }}
+              />
+              <CaregiverTextDiv>
+                <ProfileTextGray>
+                  <ProfileTextStrong>{care.userName}</ProfileTextStrong> 님
+                </ProfileTextGray>
+                <ProfileTextGray>
+                  나이
+                  <ProfileTextStrong>
+                    {care.age} 세 ({care.gender === 'M' ? '남' : care.gender === 'F' ? '여' : '성별 정보 없음'})
+                  </ProfileTextStrong>
+                </ProfileTextGray>
+              </CaregiverTextDiv>
+            </CaregiverDiv>
+
+            <CargiverButtonDiv>
+              <CareLogButton
+                onClick={() =>
+                  navigate(`/caregiverProfile/${Number(care.caregiverNo)}`, {
+                    state: { matNo: care.matNo },
+                  })
+                }
+              >
+                간병인 정보
+              </CareLogButton>
+              {activeTab === 'matched' && (
+                <ReportButton
+                  style={{ visibility: care.reviewNo ? 'hidden' : 'visible' }}
+                  onClick={() => {
+                    console.log('selected care:', care);
+                    setSelectedCaregiver(care); // 선택한 매칭 정보 저장
+                    setShowReviewModal(true); // 모달 표시
+                  }}
+                >
+                  리뷰 작성
+                </ReportButton>
+              )}
+            </CargiverButtonDiv>
+          </NewWrap>
+        ))): (
+          <EmptyMessage>매칭된 간병이 없습니다.</EmptyMessage>
+        )
+      } 
+
+      {activeTab === 'matched' ? (
+        <PageWrapper>
+          <Paging currentPage={endedCurrentPage} totalPage={endedTotalPage} chagneCurrentPage={handleEndedPageChange} />
+        </PageWrapper>
+      ) : (
+        <></>
+      )}
+    </Div>
   );
 };
 
 export default PatientCardGroup;
 
-const EmptyMessage = styled.p`
-  width: 100%;
-  text-align: center;
-  color: ${({ theme }) => theme.colors.gray[3]};
-  font-size: ${({ theme }) => theme.fontSizes.base};
-  padding: ${({ theme }) => theme.spacing[8]} 0;
+const NewWrap = styled(CargiverWrap)`
+  display: ${({ isOpen }) => (isOpen ? 'none' : 'flex')};
 `;
+
+const Div = styled.div`
+  padding-bottom: 80px; //
+  position: relative;
+`;
+
+
