+++ front/src/hooks/searchForm.js
@@ -3,11 +3,11 @@ import { matchingService } from '../api/matching';
 
 export const searchForm = () => {
 
+
+
   // 날짜 검색 함수
   const getSearchDateList =  async (patNo, page = 1 , size = 5, startDate, endDate) => {
-
     console.log("확인", startDate)
-
 
     try {
       const  data  = await matchingService.getSearchingList(patNo, startDate, endDate, page - 1, size);
@@ -17,6 +17,8 @@ export const searchForm = () => {
       console.log("실패입니다.", error);
     }
   }
+
+  
 
  
   return {
